# src/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hikma Engine Semantic Search API
  description: |
    A comprehensive semantic search API that provides powerful search capabilities over indexed codebases.
    The API leverages vector embeddings, AST parsing, git analysis, and multi-database storage to offer
    semantic similarity search, structural queries, git history searches, and hybrid multi-dimensional searches.
    
    ## Features
    - **Semantic Search**: Natural language and code similarity search using vector embeddings
    - **Structural Search**: AST-based code structure queries for functions, classes, and modules
    - **Git History Search**: Search through commit history, authors, and code evolution
    - **Hybrid Search**: Multi-dimensional search combining semantic, structural, and temporal data
    - **Comprehensive Search**: All-in-one search with facets, suggestions, and result categorization
    - **Real-time Monitoring**: Health checks, performance metrics, and error tracking
    - **Result Enhancement**: Syntax highlighting, relevance scoring, and related file discovery
    
    ## Authentication
    The API supports multiple authentication methods:
    - **API Key**: Include `X-API-Key` header with your API key
    - **JWT**: Include `Authorization: Bearer <token>` header with your JWT token
    - **Optional**: Many endpoints work without authentication for development
    
    ## Rate Limiting
    API requests are rate-limited per IP address. Default limits:
...
    
...
servers:
...
paths:
...
  /monitoring/status:
    get:
      tags:
        - Monitoring
      summary: Simple Status
      description: Simple status endpoint for load balancers
      operationId: getSimpleStatus
      responses:
        '200':
          description: Service is operational
          content:
            text/plain:
              schema:
                type: string
                enum: [OK, DEGRADED]
        '503':
          description: Service is unhealthy
          content:
            text/plain:
              schema:
                type: string
                enum: [UNHEALTHY, ERROR]
  /monitoring/readiness:
    get:
      tags:
        - Monitoring
      summary: Readiness Probe
      description: Kubernetes readiness probe endpoint
      operationId: getReadinessStatus
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ready]
        '503':
          description: Service is not ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [not ready]
                  failures:
                    type: array
                    items:
                      type: string
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {} # Allow unauthenticated access for development
...